# src/Services/Catalog/Src/Presentation/Catalog.WebApi/Internal/Variants/VariantResponse.cs
using MarketAdvanced.Catalog.Application.Common.Variants;

namespace MarketAdvanced.Catalog.WebApi.Internal.Variants;

/// <summary>Контракт GET api/internal/variants/{id} для других сервисов. Поля менять только с оглядкой на потребителей (Cart).</summary>
public sealed record VariantResponse(
    Guid Id,
    Guid ProductId,
    string ProductName,
    string Sku,
    string? Name,
    decimal Price,
    int Stock,
    bool IsActive)
{
    public static VariantResponse From(VariantDetailsResult r) => new(
        r.Id, r.ProductId, r.ProductName, r.Sku, r.Name, r.Price, r.Stock, r.IsActive);
}